[Effen] Make syncChild() support the _new_ child being null.
This is needed for cases where the UINode doesn't have a list of
children (so it doesn't go through the RenderNodeWrapper sync logic
that removes children), but it still has multiple slots, and needs to
support removing nodes from those slots. (For example, removing a
drawer from a ScaffoldContainer.)

Also, expose syncChild (it used to be private) so it can be overridden
in descendants outside fn.dart.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/1158563003