Remove users and delete //third_party/khronos headers
This stops using the GL/EGL/etc headers in //third_party/khronos on all
platforms and deletes them.  On Android, we use the EGL and OpenGL
headers from the Android NDK and define the enum values from the
following extensions if the platform does not provide them:

https://www.khronos.org/registry/egl/extensions/NV/EGL_NV_post_sub_buffer.txt
https://www.khronos.org/registry/egl/extensions/EXT/EGL_EXT_create_context_robustness.txt

On Linux we just need to know enough to talk to X in the theoretical
case that we want to swap OSMesa-produced pixels to an X window.

On Mac/iOS we're already using the headers from the relevant Frameworks.