Expose platform view ID on embedder semantics node (#13345)
This exposes platform_view_id on the embedder API's FlutterSemanticNode.

In b9261e8b52e1c6c772c60b4d99eb6a56b4a995f0 (#8055), platformViewId was
added to SemanticsNode. This field is non-zero when the SemanticsNode
represents a platform view and is typically used by embedders as a means
of identifying locations where a platform view's 'native' accessibility
tree should be injected into the platform-specific accessibility tree
constructed by the embedder.

Due to the intended use of this field, the Flutter framework is meant to
enforce that this node has a child count of zero.