Remove buggy assertion in EmbedderTest::CanPostTaskToAllNativeThreads. (#16071)
This was introduced in
https://github.com/flutter/engine/commit/66573e0c0124cf251670283cd694f9eb366d64b3.

The assertion was originally written to check that more than 4 threads were
managed by the engine (UI, Platform, GPU, IO + ConcurrentWQWorkers). However,
the assertion actually only checked the count of workers in the ConcurrentWQ. As
written, this assertion would fail on all hosts with a hardware concurrency of
less than 4. Remove the assertion. The engine threads count and its assertions
already check callbacks on workers. So this check was incorrect and redundant.