Search multiple paths when loading deferred component .so files. (#23849)
This is a partial revert of 1c4c7b6baffda57f506c1292320a89a28fd518c9

On some devices we still need to the original search paths approach
because dlopen with just the base file name doesn't work. We're
combining both approaches now, adding the base filename as the first
entry in the searchPaths.