[embedder] Ensure destruction called on present (#38078)
This is a followup to flutter/engine#38038.

In that patch, a destruction callback for textures created via the
FlutterRendererConfig callbacks (as opposed to by the FlutterCompositor
callbacks) was added and passed through to the texture info attached to
the SurfaceFrame generated in
GPUSurfaceMetalSkia::AcquireFrameFromMTLTexture (called via
GPUSurfaceMetalSkia::GetMTLTexture, which invokes the
get_next_drawable_callback), however, in order for the destruction
callback to make it to the presented Skia texture, it needs to be passed
through to the present callback here:
https://github.com/flutter/engine/blob/fdb9ace51960ab4735d297dab58b51f79945c60f/shell/gpu/gpu_surface_metal_skia.mm#LL233

which is invoked by the submit callback passed to Skia:
https://github.com/flutter/engine/blob/fdb9ace51960ab4735d297dab58b51f79945c60f/shell/gpu/gpu_surface_metal_skia.mm#LL239

The present callback is implemented in EmbedderSurface::PresentTexture,
which invokes the present callback registered in
FlutterMetalRendererConfig.present_drawable_callback. This patch ensures
that the destruction callback is passed through to Skia's present
callback so destruction occurs in the right place.

Issue: https://github.com/flutter/flutter/issues/116381