[Impeller] scales blur coverage to match rendered output (#47621)
in pursuit of: https://github.com/flutter/flutter/issues/131580

I'm starting to write tests for the new blur, but I wanted to make tests for the old blur to make sure I understood the coverage rules.  I decided to check this in separately while I work on the other blur.

Even though the goal is to delete this class eventually.  Most of this work transfers over to the new blur.

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style